Separates tck.inspection code into subpackages
