+++ tools/tck-inspection/src/main/scala/org/opencypher/tools/tck/inspection/browser/web/CSS.scala
@@ -25,7 +25,7 @@
  * described as "implementation extensions to Cypher" or as "proposed changes to
  * Cypher that are not yet approved by the openCypher community".
  */
-package org.opencypher.tools.tck.inspection
+package org.opencypher.tools.tck.inspection.browser.web
 
 import scalatags.Text.all._
 import scalatags.stylesheet.CascadingStyleSheet
