+++ tools/tck-inspection/src/main/scala/org/opencypher/tools/tck/inspection/browser/web/DiffPages.scala
@@ -25,10 +25,21 @@
  * described as "implementation extensions to Cypher" or as "proposed changes to
  * Cypher that are not yet approved by the openCypher community".
  */
-package org.opencypher.tools.tck.inspection
+package org.opencypher.tools.tck.inspection.browser.web
 
 import org.opencypher.tools.tck.api.Pickle
 import org.opencypher.tools.tck.api.Scenario
+import org.opencypher.tools.tck.inspection.collect
+import org.opencypher.tools.tck.inspection.collect.Feature
+import org.opencypher.tools.tck.inspection.collect.Group
+import org.opencypher.tools.tck.inspection.collect.ScenarioCategory
+import org.opencypher.tools.tck.inspection.collect.Total
+import org.opencypher.tools.tck.inspection.diff.ElementAdded
+import org.opencypher.tools.tck.inspection.diff.ElementRemoved
+import org.opencypher.tools.tck.inspection.diff.ElementUnchanged
+import org.opencypher.tools.tck.inspection.diff.ElementaryDiff
+import org.opencypher.tools.tck.inspection.diff.GroupDiff
+import org.opencypher.tools.tck.inspection.diff.ScenarioDiff
 import scalatags.Text
 import scalatags.Text.all._
 
@@ -115,7 +126,7 @@ case class DiffPages(diffModel: DiffModel, diffRoutes: DiffRoutes) extends PageB
         case Total => 0
         case _:ScenarioCategory => 1
         case _:Feature => 2
-        case _:Tag => 3
+        case _:collect.Tag => 3
       }
       // within each class ordered alphabetically by name
       val groupsOrdered = groupsByClasses.toSeq.sortBy(_._1).flatMap {
