+++ tools/tck-inspection/src/main/scala/org/opencypher/tools/tck/inspection/collect/Group.scala
@@ -25,7 +25,7 @@
  * described as "implementation extensions to Cypher" or as "proposed changes to
  * Cypher that are not yet approved by the openCypher community".
  */
-package org.opencypher.tools.tck.inspection
+package org.opencypher.tools.tck.inspection.collect
 
 trait Group {
   def name: String
