+++ tools/tck-inspection/src/main/scala/org/opencypher/tools/tck/inspection/collect/GroupCollection.scala
@@ -25,7 +25,7 @@
  * described as "implementation extensions to Cypher" or as "proposed changes to
  * Cypher that are not yet approved by the openCypher community".
  */
-package org.opencypher.tools.tck.inspection
+package org.opencypher.tools.tck.inspection.collect
 
 import org.opencypher.tools.tck.api.Scenario
