+++ tools/tck-inspection/src/test/resources/org/opencypher/tools/tck/inspection/browser/cli/cucumber/foo/bar/boo/Boo.feature
