+++ tools/tck-inspection/src/test/resources/org/opencypher/tools/tck/inspection/browser/cli/cucumberDiff/foo/bar/Test1.feature
