+++ tools/tck-inspection/src/test/resources/org/opencypher/tools/tck/inspection/browser/cli/cucumberDiff/foo/bar/boo/Test2.feature
