+++ tools/tck-inspection/src/test/resources/org/opencypher/tools/tck/inspection/browser/cli/cucumberDiff/foo/dummy/Dummy.feature
